requirements.txt
python-dotenv
yfinance
packaging 
duckduckgo-search
fastapi
uvicorn
groq
phidata
openai